# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Rika Events - Enterprise Event Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── INQUIRIES ─────────────────────────────────────────
...
  /rooms:
    get:
      operationId: getRooms
      tags: [rooms]
      summary: List rooms
      parameters:
        - name: venueId
          in: query
          schema:
            type: integer
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    patch:
      operationId: updateRoom
      tags: [rooms]
      summary: Update room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── VENDORS ───────────────────────────────────────────
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: eventId
          in: query
          schema:
            type: integer
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── PAYMENTS ──────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: string
        password:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        avatar:
          type: string
        status:
          type: string
        role:
          type: string

    # ── INQUIRY ───────────────────────────────────────
...
    Room:
      type: object
      required: [id, venueId, name, type, capacity, pricePerNight, status, createdAt]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        venueName:
          type: ["string", "null"]
        name:
          type: string
        type:
          type: string
          enum: [standard, deluxe, suite, presidential, hall, conference]
        capacity:
          type: integer
        pricePerNight:
          type: number
        status:
          type: string
          enum: [available, occupied, maintenance]
        floor:
          type: ["integer", "null"]
        amenities:
          type: ["string", "null"]
        guestName:
          type: ["string", "null"]
        checkIn:
          type: ["string", "null"]
        checkOut:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RoomInput:
      type: object
      required: [venueId, name, type, capacity, pricePerNight]
      properties:
        venueId:
          type: integer
        name:
          type: string
        type:
          type: string
        capacity:
          type: integer
        pricePerNight:
          type: number
        floor:
          type: integer
        amenities:
          type: string
    RoomUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        capacity:
          type: integer
        pricePerNight:
          type: number
        status:
          type: string
        guestName:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        amenities:
          type: string

    # ── VENDOR ────────────────────────────────────────
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        eventId:
          type: ["integer", "null"]
        eventTitle:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, review, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: ["string", "null"]
          format: date
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [title, status, priority]
      properties:
        title:
          type: string
        description:
          type: string
        eventId:
          type: integer
        assigneeId:
          type: integer
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
          format: date
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        assigneeId:
          type: integer
        dueDate:
          type: string
          format: date

    # ── PAYMENT ───────────────────────────────────────
...